CLI Option flags, and added the TaskfileFinder class to remove duplication
